\name{Application}
\alias{Application}
\docType{data}
\title{Application}
\description{Application data set}
\usage{data("Application")}
\format{
  The format is:
 int [1:48, 1:15] 6 9 7 5 6 7 9 9 9 4 ...
 - attr(*, "dimnames")=List of 2
  ..$ : NULL
  ..$ : chr [1:15] "FL" "APP" "AA" "LA" ...
}
\details{It is the scoring of 15 indicators on 48 interviewees}
\examples{
data(Application)
## maybe str(Application) ; plot(Application) ...
}
\keyword{datasets}
