\name{gt2013}
\alias{gt2013}
\docType{data}
\title{
Gas-Turbine CO and NOx Emission Data
}
\description{
Gas-Turbine CO and NOx Emission Data in 2013
}
\usage{data("gt2013")}
\format{
  A data frame with 7152 observations on the following 11 variables.
  \describe{
    \item{\code{AT}}{a numeric vector}
    \item{\code{AP}}{a numeric vector}
    \item{\code{AH}}{a numeric vector}
    \item{\code{AFDP}}{a numeric vector}
    \item{\code{GTEP}}{a numeric vector}
    \item{\code{TIT}}{a numeric vector}
    \item{\code{TAT}}{a numeric vector}
    \item{\code{TEY}}{a numeric vector}
    \item{\code{CDP}}{a numeric vector}
    \item{\code{CO}}{a numeric vector}
    \item{\code{NOX}}{a numeric vector}
  }
}
\details{
The dataset contains 36733 instances of 11 sensor measures aggregated over one hour, from a gas turbine located in Turkey for the purpose of studying flue gas emissions, namely CO and NOx.
}
\source{
Heysem Kaya, Department of Information and Computing Sciences, Utrecht University, 3584 CC, Utrecht, The Netherlands
}
\examples{
data(gt2013)
}
\keyword{datasets}
