\name{Iris}
\alias{Iris}
\docType{data}
\title{Iris}
\description{Iris data set}
\usage{data("Iris")}
\format{
  A data frame with 150 observations on the following 5 variables.
  \describe{
    \item{\code{Sepal.length}}{a numeric vector}
    \item{\code{Sepal.width}}{a numeric vector}
    \item{\code{Petal.length}}{a numeric vector}
    \item{\code{Petal.width}}{a numeric vector}
    \item{\code{Species}}{a character vector}
  }
}
\details{It contains 150 samples with 5 variables}
\source{Gaspar peninsula in Canada}
\examples{
data(Iris)
## maybe str(Iris) ; plot(Iris) ...
}
\keyword{datasets}
